% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{emdn_get_bbox}
\alias{emdn_get_bbox}
\alias{emdn_get_WGS84bbox}
\alias{emdn_get_nil_value}
\alias{emdn_get_band_name}
\alias{emdn_get_uom}
\alias{emdn_get_constraint}
\alias{emdn_get_grid_size}
\alias{emdn_get_resolution}
\alias{emdn_get_coverage_function}
\alias{emdn_get_temporal_extent}
\alias{emdn_get_vertical_extent}
\alias{emdn_get_dimensions_info}
\alias{emdn_get_dimensions_names}
\alias{emdn_get_dimensions_n}
\alias{emdn_get_dimension_types}
\title{Get coverage metadata from a \verb{<WCSCoverageSummary>} object.}
\usage{
emdn_get_bbox(summary)

emdn_get_WGS84bbox(summary)

emdn_get_nil_value(summary)

emdn_get_band_name(summary)

emdn_get_uom(summary)

emdn_get_constraint(summary)

emdn_get_grid_size(summary)

emdn_get_resolution(summary)

emdn_get_coverage_function(summary)

emdn_get_temporal_extent(summary)

emdn_get_vertical_extent(summary)

emdn_get_dimensions_info(
  summary,
  format = c("character", "list", "tibble"),
  include_coeffs = FALSE
)

emdn_get_dimensions_names(summary)

emdn_get_dimensions_n(summary)

emdn_get_dimension_types(summary)
}
\arguments{
\item{summary}{a \verb{<WCSCoverageSummary>} object.}

\item{format}{character string. Coverage dimension info output format.
One of \code{"character"} (default), \code{"list"} or \code{"tibble"}.}

\item{include_coeffs}{whether to include a vector of temporal or vertical
dimension coefficients (if applicable) in the coverage dimension info
\code{"list"} output format. Defaults to \code{FALSE}. Ignored for other formats.}
}
\value{
\itemize{
\item \code{emdn_get_bbox}: an object of class \code{bbox} of length 4 expressing the
boundaries coverage extent/envelope. See \code{\link[sf:st_bbox]{sf::st_bbox()}} for more details.
\item \code{emdn_get_WGS84bbox}: an object of class \code{bbox} of length 4 expressing the
boundaries coverage extent/envelope. See \code{\link[sf:st_bbox]{sf::st_bbox()}} for more details.
\item \code{emdn_get_nil_value} a numeric scalar of the value representing nil values
in a coverage.
\item \code{emdn_get_band_name} a character vector of band names.
\item \code{emdn_get_uom} a character vector of band units of measurement.
\item \code{emdn_get_constraint} a numeric vector of length 2 indicating the min and max
of the data contained in the bands of the coverage.
\item \code{emdn_get_grid_size} a numeric vector of length 2 giving the spatial size in
grid cells (pixels) of the coverage grid (ncol x nrow)
\item \code{emdn_get_resolution} a numeric vector of length 2 giving the spatial resolution
of grid cells (size in the \code{x} dimension, size in the \code{y} dimension) of a coverage. The attached attribute \code{uom} gives the units of
measurement of each dimension.
\item \code{emdn_get_coverage_function} a list with elements:
\itemize{
\item \code{sequence_rule}, character string, the function describing the sequence
rule, i.e. the relationship between the axes of data and coordinate system
axes.
\item \code{starting_point} a numeric vector of length 2, the location of the origin
of the data in the coordinate system.
}
\item \code{emdn_get_temporal_extent} if the coverage has a temporal dimension, a numeric
vector of length 2 giving the min and max values of the dimension.
Otherwise, NA.
\item \code{emdn_get_vertical_extent} if the coverage has a vertical dimension, a numeric
vector of length 2 giving the min and max values of the dimension.
Otherwise, NA.
\item \code{emdn_get_dimensions_info} output depends on \code{format} argument:
\itemize{
\item \code{character}: (default) a concatenated character string of dimension
information
\item \code{list}: a list of dimension information
\item \code{tibble}: a tibble of dimension information
}
}
}
\description{
Get coverage metadata from a \verb{<WCSCoverageSummary>} object.
}
\section{Functions}{
\itemize{
\item \code{emdn_get_bbox}: Get the bounding box (geographic extent) of a
coverage. Coordinates are given in the same Coordinate Reference System
as the coverage.

\item \code{emdn_get_WGS84bbox}: Get the bounding box (geographic extent) of a
coverage in World Geodetic System 1984 (WGS84) Coordinate Reference System
(or \code{EPSG:4326}).

\item \code{emdn_get_nil_value}: Get the value representing nil values in a
coverage.

\item \code{emdn_get_band_name}: Get the band names of a coverage.

\item \code{emdn_get_uom}: Get the units of measurement of the data contained in
the bands values of a coverage.

\item \code{emdn_get_constraint}: Get the range of values of the data contained in
the bands of the coverage.

\item \code{emdn_get_grid_size}: Get the grid sizes of a coverage

\item \code{emdn_get_resolution}: Get the resolution of a coverage

\item \code{emdn_get_coverage_function}: Get the grid function of a coverage.

\item \code{emdn_get_temporal_extent}: Get the grid function of a coverage.

\item \code{emdn_get_vertical_extent}: Get the vertical (elevation) extent of a coverage.

\item \code{emdn_get_dimensions_info}: Get information on dimensions of a coverage in
various formats. Information includes dimension label, type, unit and
range (in tibble format).

\item \code{emdn_get_dimensions_names}: Get coverage dimension names (labels).

\item \code{emdn_get_dimensions_n}: Get number of coverage dimensions.

\item \code{emdn_get_dimension_types}: Get dimensions types of a coverage.
}}

\examples{
wcs <- emdn_init_wcs_client(service = "biology")
summaries <- emdn_get_coverage_summaries_all(wcs)
summary <- summaries[[1]]
# get bbox
emdn_get_bbox(summary)
# get WGS84 bbox
emdn_get_WGS84bbox(summary)
# get the nil value of a coverage
emdn_get_nil_value(summary)
# get coverage band names
emdn_get_band_name(summary)
# get band units of measurement
emdn_get_uom(summary)
# get range of band values
emdn_get_constraint(summary)
# get coverage grid size
emdn_get_grid_size(summary)
# get coverage resolution
emdn_get_resolution(summary)
# get coverage grid function
emdn_get_coverage_function(summary)
# get the extent of the temporal dimension
emdn_get_temporal_extent(summary)
# get the extent of the vertical dimension
emdn_get_vertical_extent(summary)
# get information about coverage dimensions in various formats
emdn_get_dimensions_info(summary)
emdn_get_dimensions_info(summary, format = "list")
emdn_get_dimensions_info(summary, format = "tibble")
# get dimension names
emdn_get_dimensions_names(summary)
# get number of dimensions
emdn_get_dimensions_n(summary)
# get dimensions types
emdn_get_dimension_types(summary)
}
