% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{emdn_get_coverage_summaries}
\alias{emdn_get_coverage_summaries}
\alias{emdn_get_coverage_summaries_all}
\alias{emdn_get_coverage_ids}
\alias{emdn_has_dimension}
\alias{emdn_get_coverage_dim_coefs}
\title{Get metadata objects from a \code{WCSClient} object.}
\usage{
emdn_get_coverage_summaries(wcs, coverage_ids)

emdn_get_coverage_summaries_all(wcs)

emdn_get_coverage_ids(wcs)

emdn_has_dimension(wcs, coverage_ids, type = c("temporal", "vertical"))

emdn_get_coverage_dim_coefs(
  wcs,
  coverage_ids,
  type = c("temporal", "vertical")
)
}
\arguments{
\item{wcs}{A \code{WCSClient} R6 object, created with function \code{\link{emdn_init_wcs_client}}.}

\item{coverage_ids}{character vector of coverage IDs.}

\item{type}{a coverage dimension type. One of \code{"temporal"} or \code{"vertical"}.}
}
\value{
\itemize{
\item \code{emdn_get_coverage_summaries}: returns a list of objects of class
\verb{<WCSCoverageSummary>} for each \code{coverage_id} provided.
\item \code{emdn_get_coverage_summaries_all}: returns a list of objects of class
\verb{<WCSCoverageSummary>} for each coverage avalable through the service.
\item \code{emdn_get_coverage_ids} returns a character vector of coverage ids.
\item \code{emdn_get_coverage_dim_coefs} returns a list containing a vector of
coefficients for each coverage requested.
}

a list containing a vector of coefficients for each coverage
requested.
}
\description{
Get metadata objects from a \code{WCSClient} object. \verb{<WCSCoverageSummary>}
in particular can be used to extract further metadata about individual
coverages.
}
\section{Functions}{
\itemize{
\item \code{emdn_get_coverage_summaries()}: Get summaries for specific coverages.

\item \code{emdn_get_coverage_summaries_all()}: Get summaries for all available
coverages from a service.

\item \code{emdn_get_coverage_ids()}: Get coverage IDs for all available
coverages from a service.

\item \code{emdn_has_dimension()}: check whether a coverage has a
temporal or vertical dimension.

\item \code{emdn_get_coverage_dim_coefs()}: Get temporal or
vertical coefficients for a coverage.

}}
\examples{
\dontrun{
wcs <- emdn_init_wcs_client(service = "biology")
cov_ids <- emdn_get_coverage_ids(wcs)
cov_ids
emdn_has_dimension(wcs,
  cov_ids,
  type = "temporal"
)
emdn_has_dimension(wcs,
  cov_ids,
  type = "vertical"
)
emdn_get_coverage_summaries(wcs, cov_ids[1:2])
emdn_get_coverage_summaries_all(wcs)
emdn_get_coverage_dim_coefs(wcs,
  cov_ids[1:2],
  type = "temporal"
)
}
}
