% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{emdn_init_wcs_client}
\alias{emdn_init_wcs_client}
\title{Initialise an EMODnet WCS client}
\usage{
emdn_init_wcs_client(
  service,
  service_version = c("2.0.1", "2.1.0", "2.0.0", "1.1.1", "1.1.0"),
  logger = c("NONE", "INFO", "DEBUG")
)
}
\arguments{
\item{service}{the EMODnet OGC WCS service name.
For available services, see \code{\link[=emdn_wcs]{emdn_wcs()}}.}

\item{service_version}{the WCS service version. Defaults to "2.0.1".}

\item{logger}{character string. Level of logger: 'NONE' for no logger, 'INFO' to get ows4R logs, 'DEBUG' for all internal logs (such as as Curl details)}
}
\value{
An \code{\link[ows4R:WCSClient]{ows4R::WCSClient}} R6 object with methods for interfacing an OGC Web Coverage Service.
}
\description{
Initialise an EMODnet WCS client
}
\examples{
\dontrun{
wcs <- emdn_init_wcs_client(service = "bathymetry")
}
}
\seealso{
\code{WCSClient} in package \code{ows4R}.
}
