% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emodnet.wcs-package.R
\docType{package}
\name{emodnet.wcs-package}
\alias{emodnet.wcs}
\alias{emodnet.wcs-package}
\title{emodnet.wcs: Access EMODnet Web Coverage Service data through R}
\description{
Access and interrogate EMODnet Web Coverage Service data through R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EMODnet/emodnet.wcs}
  \item \url{https://emodnet.github.io/emodnet.wcs/}
  \item Report bugs at \url{https://github.com/EMODnet/emodnet.wcs/issues}
}

}
\author{
\strong{Maintainer}: Salvador Fernández-Bejarano \email{salvador.fernandez@vliz.be} (\href{https://orcid.org/0000-0003-0535-7677}{ORCID})

Authors:
\itemize{
  \item Anna Krystalli \email{annakrystalli@googlemail.com} (\href{https://orcid.org/0000-0002-2378-4915}{ORCID})
}

Other contributors:
\itemize{
  \item European Marine Observation Data Network (EMODnet) Biology project European Commission's Directorate - General for Maritime Affairs and Fisheries (DG MARE) \email{bio@emodnet.eu} [copyright holder]
  \item VLIZ (VLAAMS INSTITUUT VOOR DE ZEE) \email{info@vliz.be} (\href{https://ror.org/0496vr396}{ROR}) [funder]
}

}
\keyword{internal}
