% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{emdn_get_coverage_dim_coefs}
\alias{emdn_get_coverage_dim_coefs}
\title{Get temporal or vertical coefficients for a coverage}
\usage{
emdn_get_coverage_dim_coefs(
  wcs,
  coverage_id,
  type = c("temporal", "vertical")
)
}
\arguments{
\item{wcs}{A \code{WCSClient} R6 object, created with function \code{\link{emdn_init_wcs_client}}.}

\item{coverage_id}{character string. Coverage ID.}

\item{type}{character string. The dimension type for which
coefficients will be returned.}
}
\value{
a vector of coefficients.
}
\description{
Get temporal or vertical coefficients for a coverage
}
\examples{
wcs <- emdn_init_wcs_client(service =  "biology")
emdn_get_coverage_dim_coefs(wcs,
                              "Emodnetbio__ratio_large_to_small_19582016_L1_err")
}
