% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{emodnet_get_wcs_info}
\alias{emodnet_get_wcs_info}
\alias{emodnet_get_all_wcs_info}
\alias{emodnet_get_wcs_coverage_info}
\title{Get WCS available coverage information}
\usage{
emodnet_get_wcs_info(
  wcs = NULL,
  service = NULL,
  service_version = c("2.0.1", "2.1.0", "2.0.0", "1.1.1", "1.1.0"),
  logger = c("NONE", "INFO", "DEBUG")
)

emodnet_get_all_wcs_info(logger = c("NONE", "INFO", "DEBUG"))

emodnet_get_wcs_coverage_info(
  wcs = NULL,
  service = NULL,
  coverages,
  service_version = c("2.0.1", "2.1.0", "2.0.0", "1.1.1", "1.1.0"),
  logger = c("NONE", "INFO", "DEBUG")
)
}
\arguments{
\item{wcs}{A \code{WCSClient} R6 object with methods for interfacing an OGC Web Feature Service.}

\item{service}{the EMODnet OGC WCS service name.
For available services, see \code{\link[=emodnet_wcs]{emodnet_wcs()}}.}

\item{service_version}{the WCS service version. Defaults to "2.0.1".}

\item{logger}{character string. Level of logger: 'NONE' for no logger, 'INFO' to get ows4R logs, 'DEBUG' for all internal logs (such as as Curl details)}

\item{coverages}{character vector of coverage IDs.}
}
\value{
\code{emodnet_get_wcs_info} & \code{emodnet_get_wcs_info} return a list of service
level metadata, including a tibble containing coverage level metadata for each
coverage available from the service. \code{emodnet_get_wcs_coverage_info} returns a list
containing a tibble of more detailed metadata for each coverage specified. See
\strong{Details} for more information.
}
\description{
Get WCS available coverage information
}
\details{
To minimize the number of requests sent to webservices,
these functions use \code{memoise} to cache results inside the active R session.
To clear the cache, re-start R or run \code{memoise::forget(emodnet_get_wcs_info)}/\code{memoise::forget(emodnet_get_wcs_coverage_info)}.
}
\section{Functions}{
\itemize{
\item \code{emodnet_get_wcs_info}: Get info on all coverages from am EMODnet WCS service.

\item \code{emodnet_get_all_wcs_info}: Get metadata on all services and all available
coverages from each service.

\item \code{emodnet_get_wcs_coverage_info}: Get metadata for specific coverages. Requires a
\code{wcs} object as input.
}}

\examples{
emodnet_get_wcs_info(service = "biology")
# Query a wcs object
wcs <- emodnet_init_wcs_client(service = "seabed_habitats")
emodnet_get_wcs_info(wcs)
# Get detailed info for specific coverages from wcs object
coverages <- c(
  "emodnet_open_maplibrary__mediseh_cora",
  "emodnet_open_maplibrary__mediseh_posidonia"
)
emodnet_get_wcs_coverage_info(wcs = wcs, coverages = coverages)
}
