% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{emdn_get_coverage_summaries}
\alias{emdn_get_coverage_summaries}
\alias{emdn_get_coverage_summaries_all}
\alias{emdn_get_coverage_ids}
\title{Get metadata from a \code{WCSClient} object.}
\usage{
emdn_get_coverage_summaries(wcs, coverage_ids)

emdn_get_coverage_summaries_all(wcs)

emdn_get_coverage_ids(wcs)
}
\arguments{
\item{wcs}{A \code{WCSClient} R6 object, created with function \code{\link{emdn_init_wcs_client}}.}

\item{coverage_ids}{character vector of coverage IDs.}
}
\value{
\itemize{
\item \code{emdn_get_coverage_summaries}: returns a list of objects of class
\verb{<WCSCoverageSummary>} for each \code{coverage_id} provided.
\item \code{emdn_get_coverage_summaries_all}: returns a list of objects of class
\verb{<WCSCoverageSummary>} for each coverage avalable through the service.
\item \code{emdn_get_coverage_ids} returns a character vector of coverage ids.
}
}
\description{
Get metadata from a \code{WCSClient} object.
}
\section{Functions}{
\itemize{
\item \code{emdn_get_coverage_summaries}: Get summaries for specific coverages.

\item \code{emdn_get_coverage_summaries_all}: Get summaries for all available
coverages from a service.

\item \code{emdn_get_coverage_ids}: Get coverage IDs for all available
coverages from a service.
}}

\examples{
wcs <- emdn_init_wcs_client(service = "biology")
emdn_get_coverage_summaries_all(wcs)
cov_ids <- emdn_get_coverage_ids(wcs)
emdn_get_coverage_summaries(wcs, cov_ids[1:2])
}
