% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{emodnet_get_coverage_dim_coefs}
\alias{emodnet_get_coverage_dim_coefs}
\title{Get temporal or vertical coefficients for a coverage}
\usage{
emodnet_get_coverage_dim_coefs(wcs, coverage, type = c("temporal", "vertical"))
}
\arguments{
\item{wcs}{A \code{WCSClient} R6 object, created with function \code{\link{emodnet_init_wcs_client}}.}

\item{coverage}{character string. Coverage ID (name).}

\item{type}{character string. The dimension type for which
coefficients will be returned.}
}
\value{
a vector of coefficients.
}
\description{
Get temporal or vertical coefficients for a coverage
}
\examples{
wcs <- emodnet_init_wcs_client(service =  "biology")
emodnet_get_coverage_dim_coefs(wcs,
                              "Emodnetbio__ratio_large_to_small_19582016_L1_err")
}
